removed LINQ to SQL, changed Viewer folder edit

+++ PictureManagerWPF/ViewModel/FolderKeyword.cs
@@ -6,12 +6,12 @@ namespace PictureManager.ViewModel {
     public string FullPath;
     public ObservableCollection<FolderKeyword> Items { get; set; }
     public FolderKeyword Parent;
-    public List<long> FolderIdList;
+    public List<int> FolderIdList;
     public string FolderIds => string.Join(",", FolderIdList);
 
     public FolderKeyword() {
       Items = new ObservableCollection<FolderKeyword>();
-      FolderIdList = new List<long>();
+      FolderIdList = new List<int>();
     }
   }
 }
